Update styling for print view of kyc summary information

- Use media query to determine if using print media
- Update dependencies to fix vulnerability

closes #70
