+++ package.json
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "@apollo/client": "^3.8.4",
-    "@carbon/charts-react": "^1.13.1",
+    "@carbon/charts-react": "^1.13.3",
     "@carbon/grid": "^11.20.0",
     "@carbon/icons-react": "^11.27.0",
     "@carbon/react": "^1.38.0",
@@ -23,7 +23,8 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.16.0",
     "rxjs": "^7.8.1",
-    "set-value": "^4.1.0"
+    "set-value": "^4.1.0",
+    "usehooks-ts": "^2.9.1"
   },
   "devDependencies": {
     "@types/carbon__icons-react": "^11.26.1",
@@ -36,7 +37,7 @@
     "@typescript-eslint/parser": "^6.7.2",
     "@vitejs/plugin-react-swc": "^3.3.2",
     "axios": "^1.5.0",
-    "eslint": "^8.49.0",
+    "eslint": "^8.50.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
     "typescript": "^5.2.2",
