+++ src/components/KycCaseOverview/index.ts
@@ -1,0 +1,2 @@
+
+export * from './KycCaseOverview';
