+++ src/components/KycCaseResult/index.ts
@@ -1,0 +1,2 @@
+
+export * from './KycCaseResult'
