+++ src/components/index.ts
@@ -14,3 +14,5 @@ export * from './AtomSelect'
 export * from './CustomerRisk'
 export * from './NegativeNews'
 export * from './KycSummary'
+export * from './KycCaseOverview'
+export * from './KycCaseResult'
